+++ src/services/wallets/WanderWalletConnector.ts
@@ -1,14 +1,14 @@
-import { ArconnectError, WalletNotInstalledError } from '@src/utils/errors';
+import { WalletNotInstalledError, WanderError } from '@src/utils/errors';
 import eventEmitter from '@src/utils/events';
 import { PermissionType } from 'arconnect';
 import { ApiConfig } from 'arweave/node/lib/api';
 
-import { ARCONNECT_UNRESPONSIVE_ERROR } from '../../components/layout/Notifications/Notifications';
+import { WANDER_UNRESPONSIVE_ERROR } from '../../components/layout/Notifications/Notifications';
 import { ArNSWalletConnector, WALLET_TYPES } from '../../types';
 import { executeWithTimeout } from '../../utils';
 import { ArweaveTransactionID } from '../arweave/ArweaveTransactionID';
 
-export const ARCONNECT_WALLET_PERMISSIONS: PermissionType[] = [
+export const WANDER_WALLET_PERMISSIONS: PermissionType[] = [
   'ACCESS_ADDRESS',
   'ACCESS_ALL_ADDRESSES',
   'ACCESS_PUBLIC_KEY',
@@ -17,7 +17,7 @@ export const ARCONNECT_WALLET_PERMISSIONS: PermissionType[] = [
   'SIGNATURE',
 ];
 
-export class ArConnectWalletConnector implements ArNSWalletConnector {
+export class WanderWalletConnector implements ArNSWalletConnector {
   private _wallet: Window['arweaveWallet'];
   contractSigner: Window['arweaveWallet'];
   constructor() {
@@ -26,49 +26,49 @@ export class ArConnectWalletConnector implements ArNSWalletConnector {
   }
 
   // The API has been shown to be unreliable, so we call each function with a timeout
-  async safeArconnectApiExecutor<T>(fn: () => T): Promise<T> {
+  async safeWanderApiExecutor<T>(fn: () => T): Promise<T> {
     if (!this._wallet)
-      throw new WalletNotInstalledError('Arconnect is not installed.');
+      throw new WalletNotInstalledError('Wander is not installed.');
     /**
-     * This is here because occasionally arconnect injects but does not initialize internally properly,
+     * This is here because occasionally wander injects but does not initialize internally properly,
      * allowing the api to be called but then hanging.
      * This is a workaround to check that and emit appropriate errors,
-     * and to trigger the workaround workflow of reloading the page and re-initializing arconnect.
+     * and to trigger the workaround workflow of reloading the page and re-initializing wander.
      */
     const res = await executeWithTimeout(() => fn(), 3000);
 
     if (res === 'timeout') {
-      throw new Error(ARCONNECT_UNRESPONSIVE_ERROR);
+      throw new Error(WANDER_UNRESPONSIVE_ERROR);
     }
     return res as T;
   }
 
   async connect(): Promise<void> {
     if (!window.arweaveWallet) {
-      window.open('https://arconnect.io');
+      window.open('https://wander.io');
 
       return;
     }
     // confirm they have the extension installed
-    localStorage.setItem('walletType', WALLET_TYPES.ARCONNECT);
-    const permissions = await this.safeArconnectApiExecutor(
+    localStorage.setItem('walletType', WALLET_TYPES.WANDER);
+    const permissions = await this.safeWanderApiExecutor(
       this._wallet?.getPermissions,
     );
     if (
       permissions &&
-      !ARCONNECT_WALLET_PERMISSIONS.every((permission) =>
+      !WANDER_WALLET_PERMISSIONS.every((permission) =>
         permissions.includes(permission),
       )
     ) {
       // disconnect due to missing permissions, then re-connect
-      await this.safeArconnectApiExecutor(this._wallet?.disconnect);
+      await this.safeWanderApiExecutor(this._wallet?.disconnect);
     } else if (permissions) {
       return;
     }
 
     await this._wallet
       .connect(
-        ARCONNECT_WALLET_PERMISSIONS,
+        WANDER_WALLET_PERMISSIONS,
         {
           name: 'ARNS - ar.io',
         },
@@ -77,17 +77,17 @@ export class ArConnectWalletConnector implements ArNSWalletConnector {
       .catch((err) => {
         localStorage.removeItem('walletType');
         console.error(err);
-        throw new ArconnectError('User cancelled authentication.');
+        throw new WanderError('User cancelled authentication.');
       });
   }
 
   async disconnect(): Promise<void> {
     localStorage.removeItem('walletType');
-    return this.safeArconnectApiExecutor(this._wallet?.disconnect);
+    return this.safeWanderApiExecutor(this._wallet?.disconnect);
   }
 
   async getWalletAddress(): Promise<ArweaveTransactionID> {
-    return this.safeArconnectApiExecutor(() =>
+    return this.safeWanderApiExecutor(() =>
       this._wallet
         ?.getActiveAddress()
         .then((res) => new ArweaveTransactionID(res)),
@@ -95,7 +95,7 @@ export class ArConnectWalletConnector implements ArNSWalletConnector {
   }
 
   async getGatewayConfig(): Promise<ApiConfig> {
-    const config = await this.safeArconnectApiExecutor(
+    const config = await this.safeWanderApiExecutor(
       this._wallet?.getArweaveConfig,
     );
     return config as unknown as ApiConfig;
@@ -106,11 +106,11 @@ export class ArConnectWalletConnector implements ArNSWalletConnector {
     const permissions = await this._wallet.getPermissions();
     if (
       permissions &&
-      !ARCONNECT_WALLET_PERMISSIONS.every((permission) =>
+      !WANDER_WALLET_PERMISSIONS.every((permission) =>
         permissions.includes(permission),
       )
     ) {
-      const missingPermissions = ARCONNECT_WALLET_PERMISSIONS.filter(
+      const missingPermissions = WANDER_WALLET_PERMISSIONS.filter(
         (permission) => !permissions.includes(permission),
       );
       eventEmitter.emit(